Admin commands for gloss videos

+++ signbank/video/models.py
@@ -10,7 +10,7 @@ import time
 import stat
 import shutil
 
-from signbank.video.convertvideo import extract_frame, convert_video, probe_format, make_thumbnail_video, generate_image_sequence, remove_stills
+from signbank.video.convertvideo import extract_frame, convert_video, make_thumbnail_video, generate_image_sequence, remove_stills
 
 from django.core.files.storage import FileSystemStorage
 from django.contrib.auth import models as authmodels
@@ -26,6 +26,42 @@ if sys.argv[0] == 'mod_wsgi':
     from signbank.dictionary.models import *
 else:
     from signbank.dictionary.models import Gloss, Language
+
+
+def filename_matches_nme(filename):
+    filename_without_extension, ext = os.path.splitext(os.path.basename(filename))
+    return re.search(r".+-(\d+)_(nme_\d+|nme_\d+_left|nme_\d+_right)$", filename_without_extension)
+
+
+def filename_matches_perspective(filename):
+    filename_without_extension, ext = os.path.splitext(os.path.basename(filename))
+    return re.search(r".+-(\d+)_(left|right|nme_\d+_left|nme_\d+_right)$", filename_without_extension)
+
+
+def filename_matches_video(filename):
+    filename_without_extension, ext = os.path.splitext(os.path.basename(filename))
+    return re.search(r".+-(\d+)$", filename_without_extension)
+
+
+def filename_matches_backup_video(filename):
+    filename_with_extension = os.path.basename(filename)
+    return re.search(r".+-(\d+)\.(mp4|m4v|mov|webm|mkv|m2v)\.(bak\d+)$", filename_with_extension)
+
+
+def flattened_video_path(relative_path):
+    """
+    This constructs the filename to be used in the DELETED_FILES_FOLDER
+    Take apart the gloss video relative path
+    If this succeeds, prefix the filename with the dataset-specific components
+    Otherwise just return the filename
+    """
+    relative_path_folders, filename = os.path.split(relative_path)
+    m = re.search(r"^glossvideo/(.+)/(..)$", relative_path_folders)
+    if m:
+        dataset_folder = m.group(1)
+        two_char_folder = m.group(2)
+        return f"{dataset_folder}_{two_char_folder}_{filename}"
+    return filename
 
 
 class GlossVideoStorage(FileSystemStorage):
@@ -355,7 +391,7 @@ class ExampleVideo(models.Model):
         from signbank.tools import generate_still_image
         try:
             generate_still_image(self)
-        except OSError:
+        except (OSError, PermissionError):
             import sys
             print('Error generating still image', sys.exc_info())
 
@@ -377,7 +413,7 @@ class ExampleVideo(models.Model):
         small_video_path = self.small_video()
         try:
             os.unlink(self.videofile.path)
-        except OSError:
+        except (OSError, PermissionError):
             pass
 
     def reversion(self, revert=False):
@@ -503,7 +539,7 @@ class AnnotatedVideo(models.Model):
             if not only_eaf:
                 video_path = os.path.join(settings.WRITABLE_FOLDER, settings.ANNOTATEDSENTENCE_VIDEO_DIRECTORY, self.annotatedsentence.get_dataset().acronym, str(self.annotatedsentence.id))
                 shutil.rmtree(video_path)
-        except OSError:
+        except (OSError, PermissionError):
             pass
 
     def get_eaffile_name(self):
@@ -694,6 +730,8 @@ class GlossVideo(models.Model):
         """Return the URL of the small version for this video
         :param use_name: whether videofile.name should be used instead of videofile.path
         """
+        if not self.videofile:
+            return None
         small_video_path = add_small_appendix(self.videofile.path)
         if os.path.exists(small_video_path):
             if use_name:
@@ -730,7 +768,7 @@ class GlossVideo(models.Model):
         from signbank.tools import generate_still_image
         try:
             generate_still_image(self)
-        except OSError:
+        except (OSError, PermissionError):
             import sys
             print('Error generating still image', sys.exc_info())
 
@@ -751,6 +789,9 @@ class GlossVideo(models.Model):
 
         if settings.DEBUG_VIDEOS:
             print('delete_files GlossVideo: ', str(self.videofile))
+
+        if not self.videofile or not self.videofile.name:
+            return
 
         small_video_path = self.small_video()
         try:
@@ -818,8 +859,6 @@ class GlossVideo(models.Model):
         # this coercion to a string type sometimes causes special characters in the filename to be a problem
         # code has been introduced elsewhere to make sure paths are the correct encoding
         glossvideoname = self.videofile.name
-        if settings.DEBUG_VIDEOS:
-            print('__str__ GlossVideo: ', self.videofile.name)
         return glossvideoname
 
     def is_glossvideonme(self):
@@ -942,37 +981,51 @@ class GlossVideoNME(GlossVideo):
         :return:
         """
         old_path = str(self.videofile)
+        if not move_files_on_disk or not old_path:
+            return
         new_path = get_video_file_path(self, old_path, nmevideo=True, perspective='', offset=self.offset, version=0)
-        if old_path != new_path:
-            if move_files_on_disk:
-                source = os.path.join(settings.WRITABLE_FOLDER, old_path)
-                destination = os.path.join(settings.WRITABLE_FOLDER, new_path)
-                if os.path.exists(source):
-                    destination_dir = os.path.dirname(destination)
-                    if not os.path.exists(destination_dir):
-                        os.makedirs(destination_dir)
-                    if os.path.isdir(destination_dir):
-                        shutil.move(source, destination)
+        if old_path == new_path:
+            return
+        source = os.path.join(settings.WRITABLE_FOLDER, old_path)
+        destination = os.path.join(settings.WRITABLE_FOLDER, new_path)
+        if not os.path.exists(source):
+            return
+        destination_dir = os.path.dirname(destination)
+        if not os.path.exists(destination_dir):
+            os.makedirs(destination_dir)
+        if os.path.isdir(destination_dir):
+            shutil.move(source, destination)
 
-                    self.videofile.name = new_path
-                    self.save()
+        self.videofile.name = new_path
+        self.save()
 
     def delete_files(self):
         """Delete the files associated with this object"""
-        if settings.DEBUG_VIDEOS:
-            print('delete_files GlossVideoNME: ', str(self.videofile))
+        old_path = str(self.videofile)
+        if not old_path:
+            return True
+        file_system_path = os.path.join(settings.WRITABLE_FOLDER, old_path)
+        if filename_matches_nme(file_system_path) is None:
+            # this points to the normal video file, just erase the name rather than delete file
+            self.videofile.name = ""
+            self.save()
+            return True
+        if not os.path.exists(file_system_path):
+            return True
         try:
-            os.unlink(self.videofile.path)
-        except (OSError, PermissionError):
-            if settings.DEBUG_VIDEOS:
-                print('delete_files exception GlossVideo OSError, PermissionError: ', str(self.videofile))
-            pass
+            os.unlink(file_system_path)
+            return True
+        except (OSError, PermissionError) as e:
+            print(e)
+            return False
 
     def reversion(self, revert=False):
         """Delete the video file of this object"""
-        print("DELETE NME VIDEO", self.videofile.name)
-        self.delete_files()
-        self.delete()
+        status = self.delete_files()
+        if not status:
+            print("DELETE NME VIDEO FAILED: ", self.videofile.name)
+        else:
+            self.delete()
 
 
 PERSPECTIVE_CHOICES = (('left', 'Left'),
@@ -1002,6 +1055,8 @@ class GlossVideoPerspective(GlossVideo):
             return
         # other code does this too. It's a dubious way to obtain the path
         old_path = str(self.videofile)
+        if not old_path:
+            return
         new_path = get_video_file_path(self, old_path, nmevideo=False, perspective=str(self.perspective))
         if old_path == new_path:
             return
@@ -1015,36 +1070,36 @@ class GlossVideoPerspective(GlossVideo):
             if os.path.isdir(destination_dir):
                 shutil.move(source, destination)
 
-            self.videofile.name = new_path
-            self.save()
-        else:
-            # on the production server this is a problem
-            msg = "Perspective video file not found: " + source
-            print(msg)
+        self.videofile.name = new_path
+        self.save()
 
     def delete_files(self):
         """Delete the files associated with this object"""
         old_path = str(self.videofile)
+        if not old_path:
+            return True
         file_system_path = os.path.join(settings.WRITABLE_FOLDER, old_path)
-        if settings.DEBUG_VIDEOS:
-            print('perspective video delete files: ', file_system_path)
+        if filename_matches_perspective(file_system_path) is None:
+            # this points to the normal video file, just erase the name rather than delete file
+            self.videofile.name = ""
+            self.save()
+            return True
         if not os.path.exists(file_system_path):
-            # Video file not found on server
-            # on the production server this is a problem
-            msg = "Perspective video file not found: " + file_system_path
-            print(msg)
-            return
+            return True
         try:
             os.unlink(file_system_path)
-        except OSError:
-            msg = "Perspective video file could not be deleted: " + file_system_path
-            print(msg)
+            return True
+        except (PermissionError, OSError) as e:
+            print(e)
+            return False
 
     def reversion(self, revert=False):
         """Delete the video file of this object"""
-        print("DELETE Perspective VIDEO", self.videofile.name)
-        self.delete_files()
-        self.delete()
+        status = self.delete_files()
+        if not status:
+            print("DELETE Perspective VIDEO FAILED: ", self.videofile.name)
+        else:
+            self.delete()
 
 
 def move_videos_for_filter(filter, move_files_on_disk: bool=False) -> None:
@@ -1205,13 +1260,8 @@ def delete_files(sender, instance, **kwargs):
         print('delete_files pre_delete: ', str(instance))
         print('delete_files settings.DELETE_FILES_ON_GLOSSVIDEO_DELETE: ', settings.DELETE_FILES_ON_GLOSSVIDEO_DELETE)
     if hasattr(instance, 'glossvideonme'):
-        # before deleting a GlossVideoNME object, delete the files
-        instance.delete_files()
+        status = instance.delete_files()
     elif hasattr(instance, 'glossvideoperspective'):
-        # before deleting a GlossVideoPerspective object, delete the files
-        instance.delete_files()
+        status = instance.delete_files()
     elif settings.DELETE_FILES_ON_GLOSSVIDEO_DELETE:
-        # before a GlossVideo object, only delete the files if the setting is True
-        # default.py has this set to false so primary gloss video files are (never) deleted
-        # check whether this conflicts with reversion. If the file is not deleted, the object cannot be deleted
         instance.delete_files()
